Implemented the initial level LBA to PPN
The driver now implements two IOCTL commads:
1) SSD_BLKDEV_GET_LBN: The command is used to send the LBA from
driver to the user.
2) SSD_BLKDEV_SET_PPN: The command is used to receive the PPN
from the user to the driver.

A user process named testapp.c is implemented to test the
scenario.

Hence, the flow is something like this: The user-process is
launched immediately after the driver is loaded. It opens the
block device node, calls the first command and waits on it.
When an actual IO request comes to the driver, the
generated LBA will wake up the user process. The user process
will receive the LBA, convert it into an equvalent PPN and calls
the second command. Throughout this time, the IO request gets
blocked. So, once the PPN is received by the driver, the IO
request is unblocked and services the IO (read/write).

Currently, there is actually a one-one mapping of the LBA and
PPN. This has to be extended to do the mapping correctly using
an FTL in the user space.

Signed-off-by: Raghavendra Rao <raghav3276@gmail.com>