Basic implementation which just works!
This is an intermediate commit made at point when the logic just
works. Basic testing on the flow was conducted by cooking a vfat
file system on the ramdisk, which was working fine. But,
problems were faced when a heavire filesystem such as ext4 was
made to cook on it. The system just hanged.

So, work in progress..

Signed-off-by: Raghavendra Rao <raghav3276@gmail.com>