Updated the RAMDisk layout.
The disk is now divided (logically) into blocks and pages.
A page buffer is implemented which would serve as an interface
between read/write from the disk, as these operations should be
perfomed on a page size granularity.

Currently, the driver has to be registered with a
user-aplication, but the user-application doesn't
serve its need. That is, the mapping (one-one) is done
at the kernel level itself. This is because, the system
froze when communicating with the user continuously.
Still need to work out on how to solve the issue!

Signed-off-by: Raghavendra Rao Ananta <raghav3276@gmail.com>