Aggregated request process
It seems that the communication with the user is creating
problems, even the system to freeze. Hence in this implementation
the requests are gathered (of the same bio) and then sent to
the user. Still it's a bottlenext. Need to optimize it even
further! Perhaps gather the requests of the 'struct request'
itself. This is on the TODO list.

In this commit, the user-space application just has to register.
The actual mapping (one-one) is done in the kernel space itself.

Signed-off-by: Raghavendra Rao Ananta <raghav3276@gmail.com>