Optimized the request buffering.
Now, the request buffering is done per 'struct request'. This
optimized the behaviour of the SSDs.

Still, the code works pretty smooth with the mapping logic
being executed in the kernel space. But, with the
user-communication involved, the system freezes.

Possible issues: May be we are trying to sleep in interrupt
context (the request function). Need to check and fix this!

Signed-off-by: Raghavendra Rao Ananta <raghav3276@gmail.com>